==> 3_autoEncoder/2_one_hot_encoding.ipynb <==
# 2. One-Hot Encoding

import json
import pandas as pd
import numpy as np
from collections import Counter
import distutils.dir_util
import pickle
import io
import os

# 데이터 로드

# json write & load 함수 정의
def write_json(data, fname):
    def _conv(o):
        if isinstance(o, (np.int64, np.int32)):
            return int(o)
        raise TypeError

    parent = os.path.dirname(fname)
    distutils.dir_util.mkpath(parent)
    with io.open(fname, "w", encoding="utf-8") as f:
        json_str = json.dumps(data, ensure_ascii=False, default=_conv)
        f.write(json_str)
        
def load_json(fname):
    with open(fname, encoding='utf-8') as f:
        json_obj = json.load(f)

    return json_obj

# 데이터 로드
train = pd.read_json("../0_data/train.json")
test = pd.read_json("../0_data/test.json")

train.info(), test.info()

# One-Hot Encoding

> Song, tag 빈도 수에 따른 의미있는 feature 선정

# 플레이리스트 내 song id 리스트
t_song = train['songs']
# 플레이리스트 내 tag 리스트
t_tag = train['tags']
# 플레이리스트 내 song id 리스트 전체 나열
song_list = [song for ply in t_song for song in ply]
# 플레이리스트 내 tag 리스트 전체 나열
tag_list = [tag for ply in t_tag for tag in ply]

# 전체 나열 리스트 중 각 song의 개수
count_song = Counter(song_list)
# 전체 나열 리스트 중 각 tag의 개수
count_tag = Counter(tag_list)

# with open('../0_data/count_song.pkl', 'wb') as f:
#     pickle.dump(count_song, f)

# with open('../0_data/count_tag.pkl', 'wb') as f:
#     pickle.dump(count_tag, f)

len(count_song)

pd.Series(count_song).describe()

=> 곡의 빈도수 편차가 심하다. 때문에, 50% 미만 사용된 곡은 마이너 하다고 해석할 수 있으므로 해당 곡을 학습에 사용하지 않는다.

pd.Series(count_tag).describe()

=> 태그의 빈도수는 편차가 더욱 심하다. 때문에, 50% 미만 사용된 태그는 일반적인 태그가 아니라고 해석할 수 있으므로 해당 태그는 학습에 사용하지 않는다.

- count_song <= 14 인 song은 학습에서 제외
- count_tag <= 8 인 tag는 학습에서 제외

song_thr = 14
mfl={}  # meaningful
for key, value in count_song.items():
    if value >= song_thr:
        mfl[key]=value
song_len = len(mfl)
song_len

tag_thr = 8
for key, value in count_tag.items():
    if value>=tag_thr:
        mfl[key]=value
tag_len=len(mfl)-song_len
tag_len

mfl_col = list(mfl.keys())
len(mfl_col) # 유의미한 song과 tag의 총 개수 (song_len+tag_len)

with open('../0_data/mfl_col.pkl', 'wb') as f:
    pickle.dump(mfl_col, f)

with open('../0_data/mfl_col.pkl', 'rb') as f:
    mfl_col = pickle.load(f)

len(mfl_col)

> zero matrix 생성

# 전체 플레이리스트를 row, 유의미한 song과 tag를 column으로 하는 zero matrix 생성
zero_matrix = np.zeros((len(train),len(mfl_col)))
zero_matrix.shape

df_zero = pd.DataFrame(zero_matrix,columns=mfl_col,index=train['id'])
df_zero.head() # matrix >> dataframe 변환

> One-Hot Encoding

- train

# 각 플레이리스트에 해당하는 곡, 태그 좌표에 1 부여

def create_onehot(df,column_name):
    zero_matrix=np.zeros((len(df),len(column_name)))
    zero_df=pd.DataFrame(zero_matrix,columns=column_name,index=df['id'])
    for i in range(len(df)):
        for j in df.iloc[i,0]+df.iloc[i,3] :
            if j in column_name:
                zero_df.iloc[i,column_name.index(j)]=1
    return zero_df

# create_onehot 함수 활용 train data >> onehot encoding
train_onehot = create_onehot(train, mfl_col)
train_onehot.shape

with open('../0_data/train_onehot.pkl', 'wb') as f:
    pickle.dump(train_onehot, f)

with open('../0_data/train_onehot.pkl', 'rb') as f:
    train_onehot = pickle.load(f)

train_onehot.shape

- test

# create_onehot 함수 활용 test data >> onehot encoding
test_onehot = create_onehot(test, mfl_col)
test_onehot.shape

with open('../0_data/test_onehot.pkl', 'wb') as f:
    pickle.dump(test_onehot, f)

with open('../0_data/test_onehot.pkl', 'rb') as f:
    test_onehot = pickle.load(f)

test_onehot.shape

---
